requirements.txt
plotly
chart-studio
matplotlib
scipy
numpy
Pillow
numpy-stl